21910 - Publishing time should not be part of the request time
In the past publishing messages was in a two second timeout which was
part of the request timeout. This lead to scenarios on slow networks or
networks with a very large amount of hosts publishing would fail
unexpectedly.

Publishing is now contained in its own timeout. The --publishing-timeout
cli flag and publishing_timeout client config options have also been
added which can now be used to reduce or extend the time allowed for
publishing. Defaults to 10 seconds.

Conflicts:
	lib/mcollective/config.rb
	lib/mcollective/rpc/client.rb
	spec/unit/config_spec.rb
	spec/unit/rpc/client_spec.rb

BACKPORT to 2.2.1