maint: add rubocop rules for style enforcement
We have a slightly inconsistent style across the codebase.

This starts the process of tidying that up by adding rubocop configurations and
disabling the style issues it reports.  Later we can debate each style issue,
and if we agree with the wider ruby community style guide as crystalised in
rubocop, remove the exceptions and fix the style issues.