MCO-100 specify stomp '~> 1.3.2'
We suggest the newest avaiable stomp gem in the 1.2.x or 1.3.x series.  1.3.0
and 1.3.1 won't work because of incompatibilty introduced and then fixed in
their logging layer refactor.

Practically we can only specify '~> 1.3.2' because there seems to be no way of
expressing >= 1.2.10 is sufficient but >= 1.3.2 will do too.   The stompgem
project follows semver, so it's very unlikely that there will be a return to
the 1.2.x release series, so it's safe enough to pessimise on '~> 1.3.2' though
it may force an upgrade in some installs.